moved entry list to subject page
