+++ src/pages/subjects/[subject]/entry.tsx
@@ -5,28 +5,12 @@ import Button from "../../../ui/Button";
 import Heading from "../../../ui/Heading";
 import ButtonContainer from "../../../ui/ButtonContainer";
 import MainContent from "../../../ui/MainContent";
-import { trpc } from "../../../utils/trpc";
-import { createServerSideHelpers } from "@trpc/react-query/server";
-import { appRouter } from "../../../server/trpc/router/_app";
-import { prisma } from "../../../server/db/client";
-import superjson from "superjson";
-import type { GetStaticPaths, GetStaticPropsContext } from "next";
 
-const Entry: NextPage<{ subject: string }> = ({ subject }) => {
-  console.log(subject);
-
-  const { data } = trpc.entry.getEntries.useQuery({
-    subjectName: subject,
-  });
-
-  if (!data) return <div>404 Not Found</div>;
-
+const Entry: NextPage = () => {
   return (
     <Layout page="New Entry">
       <Heading>New Entry</Heading>
-      <MainContent>
-        <TextArea />
-      </MainContent>
+      <MainContent></MainContent>
       <ButtonContainer>
         <Button intent="cancel" link="/">
           Back
@@ -53,33 +37,5 @@ const TextArea = () => {
     </div>
   );
 };
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  return {
-    paths: [],
-    fallback: "blocking",
-  };
-};
-
-export async function getStaticProps(
-  context: GetStaticPropsContext<{ subject: string }>
-) {
-  const ssg = await createServerSideHelpers({
-    router: appRouter,
-    ctx: { prisma, session: null },
-    transformer: superjson,
-  });
-
-  const subject = context.params?.subject as string;
-
-  await ssg.entry.getEntries.prefetch({ subjectName: subject });
-
-  return {
-    props: {
-      trpcState: ssg.dehydrate(),
-      subject: subject,
-    },
-  };
-}
 
 export default Entry;
