+++ src/server/trpc/router/entry.ts
@@ -1,4 +1,3 @@
-import { TRPCError } from "@trpc/server";
 import { router, protectedProcedure } from "../trpc";
 import z from "zod";
